added post function to send form data to back end for email
